requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
httpx>=0.24.0,<0.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0,<=0.21.1
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
